backend/requirements.txt
Flask>=2.3.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
Flask-JWT-Extended>=4.6.0
Flask-SocketIO>=5.3.0
python-socketio[client]>=5.11.0
eventlet>=0.36.0
deepface>=0.0.79
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Werkzeug>=3.0.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.2
nltk>=3.8.1
scikit-learn>=1.3.2
requests>=2.31.0
psycopg2-binary>=2.9.0
gunicorn>=21.2.0
tensorflow>=2.13.0
tf-keras>=2.13.0
keras>=2.13.0
pandas>=2.0.0
matplotlib>=3.7.0
tqdm>=4.65.0

